+++ wc_dart_framework/example/lib/bloc_example/bloc_example_state.dart
@@ -20,6 +20,7 @@ abstract class BlocExampleState
 
   @BlocHydratedField()
   @BlocUpdateField()
+  @BlocListenField()
   String? get value1;
 
   @BlocHydratedField()