+++ wc_dart_framework/lib/wc_dart_framework.dart
@@ -6,3 +6,4 @@ export 'src/annotations/asset.dart';
 export 'src/annotations/bloc.dart';
 export 'src/annotations/enum.dart';
 export 'src/utils/logging_utils.dart';
+export 'src/utils/bloc_basic_listener.dart';